Remove two sources of reading material
Both resources are very far from beginner friendly and clearly suppose previous coding experiences. 
Example of a sentence used in these resources:

"javaScript contains a standard library of objects, such as Array, Date, and Math, and a core set of language elements such as operators, control structures, and statements. Core JavaScript can be extended for a variety of purposes by supplementing it with additional objects"

As the level of English of some of our students is limited, and many have no background in programming, I think this resource is mostly just having students read information that they are not yet ready to process, and mostly will leave them confused. There are much better resources (that we have written ourselves) that explain it much better, and that we don't link to in the reading material of this week. (I have created a seperate pull request for this.