Alternative readings, instead of MDM and SpeakinJS
As replacement to MDM and speakingjs webpage(too compicated), I thought it would be a good idea to replace them with our own documentation on the topics we discuss that week. These links are now considered Bonus material, but I think it's good that are obliged to read up on the topics that will be discussed in class, so that the teachings actually stick better with the students.