Changed links of recourses, more beginner friendly
From the students I've gotten the signal that MDM is not helping them learn to code, as it is written for experienced developers and mostly just confuses them.  
Also the book "a simpler way to learn javaScript" is getting quite outdated (using var only e.g), and is a bit of a pain to read as it is a pdf file with hundreds of pages.

Many students have expressed to have had great help from the website javaScript.info, and I think it's a great webpage that is beginner friendly, but still manages to explain concepts well end in depth.