replace another link that is too hard to read
replaced 2 links ( from eloquent javaScript) that is too hard to read for beginners / non native english speakers. 
the first I replaced it with our internal Scope documentation, that currently wasn't part of the reading (and is talked about in Eloquent JS) 
the second I replaced with a clear video on expressions vs statements. I think in general it's good to be very specific about what each reading is exactly about, and what you should get understand out of it.